requirements.txt
google-generativeai
google-api-python-client 
google-auth 
google-auth-oauthlib 
python-dotenv
python-dateutil
parsedatetime
google-auth-httplib2
pytz

